@@ -4,14 +4,13 @@ part of 'element.dart';
 class FamilyElement extends GedcomElement {
   /// Constructor of the FamilyElement
   FamilyElement({
-    @required int level,
-    String pointer,
-    String value,
-    List<GedcomElement> children,
-    GedcomElement parent,
-    String crlf = '\n',
-  })  : assert(level != null, 'Level is required'),
-        super(
+    required int level,
+    String? pointer,
+    String? value,
+    List<GedcomElement>? children,
+    GedcomElement? parent,
+    String? crlf = '\n',
+  }) : super(
           level: level,
           tag: GEDCOM_TAG_FAMILY,
           pointer: pointer,
@@ -27,13 +26,12 @@ class FamilyElement extends GedcomElement {
   /// Returns the husband/father of the family
   ///
   /// Returns null if not found
-  IndividualElement getHusband(Map<String, GedcomElement> elementsMap) {
-    final husbandId = children.firstWhere(
-        (element) => element.tag == GEDCOM_TAG_HUSBAND,
-        orElse: () => null);
+  IndividualElement? getHusband(Map<String, GedcomElement> elementsMap) {
+    final husbandId = children
+        .firstWhereOrNull((element) => element.tag == GEDCOM_TAG_HUSBAND);
     if (husbandId != null) {
-      final husband = elementsMap[husbandId.value];
-      return husband;
+      final husband = elementsMap[husbandId.value!];
+      return husband as IndividualElement?;
     }
     return null;
   }
@@ -41,13 +39,12 @@ class FamilyElement extends GedcomElement {
   /// Returns the wife/mother of the family
   ///
   /// Returns null if not found
-  IndividualElement getWife(Map<String, GedcomElement> elementsMap) {
-    final wifeId = children.firstWhere(
-        (element) => element.tag == GEDCOM_TAG_WIFE,
-        orElse: () => null);
+  IndividualElement? getWife(Map<String, GedcomElement> elementsMap) {
+    final wifeId =
+        children.firstWhereOrNull((element) => element.tag == GEDCOM_TAG_WIFE);
     if (wifeId != null) {
-      final wife = elementsMap[wifeId.value];
-      return wife;
+      final wife = elementsMap[wifeId.value!];
+      return wife as IndividualElement?;
     }
     return null;
   }
@@ -55,12 +52,12 @@ class FamilyElement extends GedcomElement {
   /// Returns the children sorted by the birth year if available
   ///
   /// Returns empty list if no children found
-  List<IndividualElement> getChildren(Map<String, GedcomElement> elementsMap) {
-    final childrenElements = <IndividualElement>[];
+  List<IndividualElement?> getChildren(Map<String, GedcomElement> elementsMap) {
+    final childrenElements = <IndividualElement?>[];
     final childrenIds =
         children.where((element) => element.tag == GEDCOM_TAG_CHILD).toList();
     for (final child in childrenIds) {
-      childrenElements.add(elementsMap[child.value]);
+      childrenElements.add(elementsMap[child.value!] as IndividualElement?);
     }
     childrenElements.sort();
     return childrenElements;
@@ -69,13 +66,13 @@ class FamilyElement extends GedcomElement {
   /// Returns copy of the element
   @override
   FamilyElement copyWith({
-    int level,
-    String pointer,
-    String tag,
-    String value,
-    List<GedcomElement> children,
-    GedcomElement parent,
-    String crlf,
+    int? level,
+    String? pointer,
+    String? tag,
+    String? value,
+    List<GedcomElement>? children,
+    GedcomElement? parent,
+    String? crlf,
   }) {
     return FamilyElement(
       level: level ?? this.level,
